add market view + mimir endpoint graphql requests

+++ 9cWebMarket.Frontend/Models/Api/SkillModel.cs
@@ -10,6 +10,6 @@ public record SkillModel(
     [property: JsonPropertyName("elementalType")] int ElementalType,
     [property: JsonPropertyName("referencedStatType")] int ReferencedStatType,
     [property: JsonPropertyName("cooldown")] int Cooldown,
-    [property: JsonPropertyName("power")] int Power,
+    [property: JsonPropertyName("power")] long Power,
     [property: JsonPropertyName("skillCategory")] int SkillCategory
 );
